+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2024.08.02`.
+ * Upgraded stable Flutter analysis SDK to `3.22.3`.
+ * Upgraded dartdoc to `8.0.12`.
 
 ## `20240718t101100-all`
  * Bumped runtimeVersion to `2024.07.18`.